fixed CSS class names
