+++ src/components/artists/artist.css
@@ -51,19 +51,15 @@
     justify-content: space-between;
     border-bottom: 1px solid rgb(199, 199, 199);
 }
-input, select {
+
+.profile__input, .profile__select .profile__textarea {
     background-color: transparent;
     border: none;
     width: fit-content;
     height: auto
 }
 
-.firstname {
-    width: auto;
-    margin: 0
-}
-
-textarea {
+.profile__textarea {
     min-height: 20rem;
     width: 100%;
     white-space: pre-wrap;
