+++ src/components/nav/NavBar.css
@@ -1,27 +1,27 @@
 .navbar { 
-    color: #ffffff; 
-    background-color: #131313;
+    color: #FB9970; 
+    background-color: #FB9970;
  }
  
  .navbar-menu img {
     max-height: 3rem;
-    color: #ffffff; 
-    background-color: #131313;
+    color: #000000; 
+    background-color: #FB9970;
  }
  
  .navbar-brand, .navbar-item, .navbar-burger, .navbar-menu { 
     font-size: 15px;
-    background-color: #131313;
+    background-color: #FB9970;
  }
  
  .navbar-item, .navbar-link { 
-    color: #cccccc;
-    background-color: #131313;
+    color: #000000;
+    background-color: #FB9970;
     padding: 0px; 
  }
  
  .navbar-item, .navbar-menu.is-active  { 
-    background-color: #131313; 
+    background-color: #FB9970; 
  }
  
  .navbar-end { 
@@ -34,11 +34,11 @@
  
  .padding { 
     padding: 45px 10px; 
-    color: #3cffd0;
+    color: #575757;
  }
  
  .rareHeader { 
-    color: #ffffff;
+    color: #343434;
     font-weight: bolder; 
     font-size: 25px; 
     margin-left: 25px;
