+++ src/views/ApplicationViews.js
@@ -9,7 +9,6 @@ import { AddPieceForm, PieceForms } from "../components/art/AddPieceForm"
 import { EditPieceForm } from "../components/art/EditPieceForm"
 import { NewFile } from "../components/art/NewFile"
 import { Profile } from "../components/artists/Profile"
-import { EditProfile } from "../components/artists/EditProfile"
 
 export const ApplicationViews = ({ token, setToken }) => {
   return (
@@ -33,7 +32,6 @@ export const ApplicationViews = ({ token, setToken }) => {
         </Route>
         <Route path="/profile" >
             <Route index element={<Profile/>} />
-            <Route path="edit" element={<EditProfile />} />
         </Route>
         <Route path="/testing" element={<NewFile />} />
         <Route element={<Authorized token={token} />}>
